Persist highlighting on color scheme change
Well-written Vim color schemes run the command :highlight clear to
remove the highlight groups set by a previous color scheme.
Unfortunately, this also removes the internal highlight group set by
vim-exchange, effectively disabling vim-exchange's highlighting, and
running :XchangeHighlightToggle does not reenable the plugin's
highlighting.

In this commit, this is fixed by adding an autocmd to link the
_exchange_region highlight group back to ExchangeRegion when the color
scheme is changed and the highlighting is enabled.